backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.0
dj-database-url>=2.1.0
python-dotenv>=1.0.0
Pillow>=10.2.0
django-storages>=1.14.2
boto3>=1.34.0
requests>=2.31.0
httpx>=0.27.0
stripe>=7.15.0